requirements.txt
facebook-sdk
python-social-auth
restless
PyJWT